@@ -23,7 +23,7 @@
     "issues": "https://github.com/phar-io/version/issues"
   },
   "require": {
-    "php": "^5.6 || ^7.0"
+    "php": "^7.2"
   },
   "autoload": {
     "classmap": [
