@@ -1,36 +1,23 @@
-<?php
-
+<?php declare(strict_types = 1);
 namespace PharIo\Version;
 
 class VersionConstraintValue {
-    /**
-     * @var VersionNumber
-     */
+    /** @var VersionNumber */
     private $major;
 
-    /**
-     * @var VersionNumber
-     */
+    /** @var VersionNumber */
     private $minor;
 
-    /**
-     * @var VersionNumber
-     */
+    /** @var VersionNumber */
     private $patch;
 
-    /**
-     * @var string
-     */
+    /** @var string */
     private $label = '';
 
-    /**
-     * @var string
-     */
+    /** @var string */
     private $buildMetaData = '';
 
-    /**
-     * @var string
-     */
+    /** @var string */
     private $versionString = '';
 
     /**
@@ -42,60 +29,42 @@ class VersionConstraintValue {
         $this->parseVersion($versionString);
     }
 
-    /**
-     * @return string
-     */
-    public function getLabel() {
+    public function getLabel(): string {
         return $this->label;
     }
 
-    /**
-     * @return string
-     */
-    public function getBuildMetaData() {
+    public function getBuildMetaData(): string {
         return $this->buildMetaData;
     }
 
-    /**
-     * @return string
-     */
-    public function getVersionString() {
+    public function getVersionString(): string {
         return $this->versionString;
     }
 
-    /**
-     * @return VersionNumber
-     */
-    public function getMajor() {
+    public function getMajor(): VersionNumber {
         return $this->major;
     }
 
-    /**
-     * @return VersionNumber
-     */
-    public function getMinor() {
+    public function getMinor(): VersionNumber {
         return $this->minor;
     }
 
-    /**
-     * @return VersionNumber
-     */
-    public function getPatch() {
+    public function getPatch(): VersionNumber {
         return $this->patch;
     }
 
     /**
      * @param $versionString
      */
-    private function parseVersion($versionString) {
+    private function parseVersion($versionString): void {
         $this->extractBuildMetaData($versionString);
         $this->extractLabel($versionString);
 
-        $versionSegments = explode('.', $versionString);
-        $this->major = new VersionNumber($versionSegments[0]);
+        $versionSegments = \explode('.', $versionString);
+        $this->major     = new VersionNumber($versionSegments[0]);
 
-        $minorValue = isset($versionSegments[1]) ? $versionSegments[1] : null;
-        $patchValue = isset($versionSegments[2]) ? $versionSegments[2] : null;
+        $minorValue = $versionSegments[1] ?? null;
+        $patchValue = $versionSegments[2] ?? null;
 
         $this->minor = new VersionNumber($minorValue);
         $this->patch = new VersionNumber($patchValue);
@@ -104,20 +73,20 @@ class VersionConstraintValue {
     /**
      * @param string $versionString
      */
-    private function extractBuildMetaData(&$versionString) {
-        if (preg_match('/\+(.*)/', $versionString, $matches) == 1) {
+    private function extractBuildMetaData(&$versionString): void {
+        if (\preg_match('/\+(.*)/', $versionString, $matches) == 1) {
             $this->buildMetaData = $matches[1];
-            $versionString = str_replace($matches[0], '', $versionString);
+            $versionString       = \str_replace($matches[0], '', $versionString);
         }
     }
 
     /**
      * @param string $versionString
      */
-    private function extractLabel(&$versionString) {
-        if (preg_match('/\-(.*)/', $versionString, $matches) == 1) {
-            $this->label = $matches[1];
-            $versionString = str_replace($matches[0], '', $versionString);
+    private function extractLabel(&$versionString): void {
+        if (\preg_match('/\-(.*)/', $versionString, $matches) == 1) {
+            $this->label   = $matches[1];
+            $versionString = \str_replace($matches[0], '', $versionString);
         }
     }
 }
