@@ -1,32 +1,25 @@
-<?php
-
+<?php declare(strict_types = 1);
 namespace PharIo\Version;
 
 class PreReleaseSuffix {
     private $valueScoreMap = [
-        'dev' => 0,
-        'a' => 1,
+        'dev'   => 0,
+        'a'     => 1,
         'alpha' => 1,
-        'b' => 2,
-        'beta' => 2,
-        'rc' => 3,
-        'p' => 4,
+        'b'     => 2,
+        'beta'  => 2,
+        'rc'    => 3,
+        'p'     => 4,
         'patch' => 4,
     ];
 
-    /**
-     * @var string
-     */
+    /** @var string */
     private $value;
 
-    /**
-     * @var int
-     */
+    /** @var int */
     private $valueScore;
 
-    /**
-     * @var int
-     */
+    /** @var int */
     private $number = 0;
 
     /**
@@ -36,26 +29,15 @@ class PreReleaseSuffix {
         $this->parseValue($value);
     }
 
-    /**
-     * @return string
-     */
-    public function getValue() {
+    public function getValue(): string {
         return $this->value;
     }
 
-    /**
-     * @return int|null
-     */
-    public function getNumber() {
+    public function getNumber(): ?int {
         return $this->number;
     }
 
-    /**
-     * @param PreReleaseSuffix $suffix
-     *
-     * @return bool
-     */
-    public function isGreaterThan(PreReleaseSuffix $suffix) {
+    public function isGreaterThan(PreReleaseSuffix $suffix): bool {
         if ($this->valueScore > $suffix->valueScore) {
             return true;
         }
@@ -69,24 +51,24 @@ class PreReleaseSuffix {
 
     /**
      * @param $value
-     *
-     * @return int
      */
-    private function mapValueToScore($value) {
-        if (array_key_exists($value, $this->valueScoreMap)) {
+    private function mapValueToScore($value): int {
+        if (\array_key_exists($value, $this->valueScoreMap)) {
             return $this->valueScoreMap[$value];
         }
 
         return 0;
     }
 
-    private function parseValue($value) {
+    private function parseValue($value): void {
         $regex = '/-?(dev|beta|b|rc|alpha|a|patch|p)\.?(\d*).*$/i';
-        if (preg_match($regex, $value, $matches) !== 1) {
-            throw new InvalidPreReleaseSuffixException(sprintf('Invalid label %s', $value));
+
+        if (\preg_match($regex, $value, $matches) !== 1) {
+            throw new InvalidPreReleaseSuffixException(\sprintf('Invalid label %s', $value));
         }
 
         $this->value = $matches[1];
+
         if (isset($matches[2])) {
             $this->number = (int)$matches[2];
         }
