@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types = 1);
 /*
  * This file is part of PharIo\Version.
  *
@@ -7,33 +7,22 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-
 namespace PharIo\Version;
 
 class Version {
-    /**
-     * @var VersionNumber
-     */
+    /** @var VersionNumber */
     private $major;
 
-    /**
-     * @var VersionNumber
-     */
+    /** @var VersionNumber */
     private $minor;
 
-    /**
-     * @var VersionNumber
-     */
+    /** @var VersionNumber */
     private $patch;
 
-    /**
-     * @var PreReleaseSuffix
-     */
+    /** @var PreReleaseSuffix */
     private $preReleaseSuffix;
 
-    /**
-     * @var string
-     */
+    /** @var string */
     private $versionString = '';
 
     /**
@@ -45,33 +34,19 @@ class Version {
         $this->versionString = $versionString;
     }
 
-    /**
-     * @return PreReleaseSuffix
-     */
-    public function getPreReleaseSuffix() {
+    public function getPreReleaseSuffix(): PreReleaseSuffix {
         return $this->preReleaseSuffix;
     }
 
-    /**
-     * @return string
-     */
-    public function getVersionString() {
+    public function getVersionString(): string {
         return $this->versionString;
     }
 
-    /**
-     * @return bool
-     */
-    public function hasPreReleaseSuffix() {
+    public function hasPreReleaseSuffix(): bool {
         return $this->preReleaseSuffix !== null;
     }
 
-    /**
-     * @param Version $version
-     *
-     * @return bool
-     */
-    public function isGreaterThan(Version $version) {
+    public function isGreaterThan(Version $version): bool {
         if ($version->getMajor()->getValue() > $this->getMajor()->getValue()) {
             return false;
         }
@@ -111,31 +86,19 @@ class Version {
         return $this->getPreReleaseSuffix()->isGreaterThan($version->getPreReleaseSuffix());
     }
 
-    /**
-     * @return VersionNumber
-     */
-    public function getMajor() {
+    public function getMajor(): VersionNumber {
         return $this->major;
     }
 
-    /**
-     * @return VersionNumber
-     */
-    public function getMinor() {
+    public function getMinor(): VersionNumber {
         return $this->minor;
     }
 
-    /**
-     * @return VersionNumber
-     */
-    public function getPatch() {
+    public function getPatch(): VersionNumber {
         return $this->patch;
     }
 
-    /**
-     * @param array $matches
-     */
-    private function parseVersion(array $matches) {
+    private function parseVersion(array $matches): void {
         $this->major = new VersionNumber($matches['Major']);
         $this->minor = new VersionNumber($matches['Minor']);
         $this->patch = isset($matches['Patch']) ? new VersionNumber($matches['Patch']) : new VersionNumber(null);
@@ -150,7 +113,7 @@ class Version {
      *
      * @throws InvalidVersionException
      */
-    private function ensureVersionStringIsValid($version) {
+    private function ensureVersionStringIsValid($version): void {
         $regex = '/^v?
             (?<Major>(0|(?:[1-9][0-9]*)))
             \\.
@@ -164,9 +127,9 @@ class Version {
             )?       
         $/x';
 
-        if (preg_match($regex, $version, $matches) !== 1) {
+        if (\preg_match($regex, $version, $matches) !== 1) {
             throw new InvalidVersionException(
-                sprintf("Version string '%s' does not follow SemVer semantics", $version)
+                \sprintf("Version string '%s' does not follow SemVer semantics", $version)
             );
         }
 
