@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types = 1);
 /*
  * This file is part of PharIo\Version.
  *
@@ -7,35 +7,23 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-
 namespace PharIo\Version;
 
 class VersionNumber {
-    /**
-     * @var int
-     */
+    /** @var int */
     private $value;
 
-    /**
-     * @param mixed $value
-     */
     public function __construct($value) {
-        if (is_numeric($value)) {
+        if (\is_numeric($value)) {
             $this->value = $value;
         }
     }
 
-    /**
-     * @return bool
-     */
-    public function isAny() {
+    public function isAny(): bool {
         return $this->value === null;
     }
 
-    /**
-     * @return int
-     */
-    public function getValue() {
+    public function getValue(): int {
         return $this->value;
     }
 }
