@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types = 1);
 /*
  * This file is part of PharIo\Version.
  *
@@ -7,26 +7,17 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-
 namespace PharIo\Version;
 
 abstract class AbstractVersionConstraint implements VersionConstraint {
-    /**
-     * @var string
-     */
-    private $originalValue = '';
+    /** @var string */
+    private $originalValue;
 
-    /**
-     * @param string $originalValue
-     */
-    public function __construct($originalValue) {
+    public function __construct(string $originalValue) {
         $this->originalValue = $originalValue;
     }
 
-    /**
-     * @return string
-     */
-    public function asString() {
+    public function asString(): string {
         return $this->originalValue;
     }
 }
