@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types = 1);
 /*
  * This file is part of PharIo\Version.
  *
@@ -7,23 +7,14 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-
 namespace PharIo\Version;
 
 class AnyVersionConstraint implements VersionConstraint {
-    /**
-     * @param Version $version
-     *
-     * @return bool
-     */
-    public function complies(Version $version) {
+    public function complies(Version $version): bool {
         return true;
     }
 
-    /**
-     * @return string
-     */
-    public function asString() {
+    public function asString(): string {
         return '*';
     }
 }
