@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types = 1);
 /*
  * This file is part of PharIo\Version.
  *
@@ -7,18 +7,14 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-
 namespace PharIo\Version;
 
 class GreaterThanOrEqualToVersionConstraint extends AbstractVersionConstraint {
-    /**
-     * @var Version
-     */
+    /** @var Version */
     private $minimalVersion;
 
     /**
      * @param string $originalValue
-     * @param Version $minimalVersion
      */
     public function __construct($originalValue, Version $minimalVersion) {
         parent::__construct($originalValue);
@@ -26,12 +22,7 @@ class GreaterThanOrEqualToVersionConstraint extends AbstractVersionConstraint {
         $this->minimalVersion = $minimalVersion;
     }
 
-    /**
-     * @param Version $version
-     *
-     * @return bool
-     */
-    public function complies(Version $version) {
+    public function complies(Version $version): bool {
         return $version->getVersionString() == $this->minimalVersion->getVersionString()
             || $version->isGreaterThan($this->minimalVersion);
     }
