@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types = 1);
 /*
  * This file is part of PharIo\Version.
  *
@@ -7,24 +7,19 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-
 namespace PharIo\Version;
 
 class SpecificMajorAndMinorVersionConstraint extends AbstractVersionConstraint {
-    /**
-     * @var int
-     */
+    /** @var int */
     private $major = 0;
 
-    /**
-     * @var int
-     */
+    /** @var int */
     private $minor = 0;
 
     /**
      * @param string $originalValue
-     * @param int $major
-     * @param int $minor
+     * @param int    $major
+     * @param int    $minor
      */
     public function __construct($originalValue, $major, $minor) {
         parent::__construct($originalValue);
@@ -33,12 +28,7 @@ class SpecificMajorAndMinorVersionConstraint extends AbstractVersionConstraint {
         $this->minor = $minor;
     }
 
-    /**
-     * @param Version $version
-     *
-     * @return bool
-     */
-    public function complies(Version $version) {
+    public function complies(Version $version): bool {
         if ($version->getMajor()->getValue() != $this->major) {
             return false;
         }
