@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types = 1);
 /*
  * This file is part of PharIo\Version.
  *
@@ -7,18 +7,15 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-
 namespace PharIo\Version;
 
 class SpecificMajorVersionConstraint extends AbstractVersionConstraint {
-    /**
-     * @var int
-     */
+    /** @var int */
     private $major = 0;
 
     /**
      * @param string $originalValue
-     * @param int $major
+     * @param int    $major
      */
     public function __construct($originalValue, $major) {
         parent::__construct($originalValue);
@@ -26,12 +23,7 @@ class SpecificMajorVersionConstraint extends AbstractVersionConstraint {
         $this->major = $major;
     }
 
-    /**
-     * @param Version $version
-     *
-     * @return bool
-     */
-    public function complies(Version $version) {
+    public function complies(Version $version): bool {
         return $version->getMajor()->getValue() == $this->major;
     }
 }
