@@ -1,7 +1,5 @@
-<?php
-
+<?php declare(strict_types = 1);
 namespace PharIo\Version;
 
 class InvalidPreReleaseSuffixException extends \Exception implements Exception {
-
 }
