@@ -1,5 +1,4 @@
-<?php
-
+<?php declare(strict_types = 1);
 namespace PharIo\Version;
 
 class InvalidVersionException extends \InvalidArgumentException implements Exception {
