@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types = 1);
 /*
  * This file is part of PharIo\Version.
  *
@@ -7,7 +7,6 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-
 namespace PharIo\Version;
 
 use PHPUnit\Framework\TestCase;
@@ -16,7 +15,7 @@ use PHPUnit\Framework\TestCase;
  * @covers \PharIo\Version\AbstractVersionConstraint
  */
 class AbstractVersionConstraintTest extends TestCase {
-    public function testAsString() {
+    public function testAsString(): void {
         /** @var AbstractVersionConstraint|\PHPUnit_Framework_MockObject_MockObject $constraint */
         $constraint = $this->getMockForAbstractClass(AbstractVersionConstraint::class, ['foo']);
 
